pyproject.toml
[project]
name = "game-night-decider"
version = "0.1.0"
description = "Telegram bot to decide what to play on game night"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
authors = [
    {name = "João Caetano", email = "joaomartinscaetano@proton.me"},
]
dependencies = [
    "httpx>=0.28.1",
    "lxml>=6.0.2",
    "python-dotenv>=1.2.1",
    "python-telegram-bot[job-queue]>=21.10",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "aiosqlite>=0.20.0",
    "pydantic>=2.10.3",
    "pydantic-settings>=2.7.0",
]

[tool.semantic_release]
version_toml = ["pyproject.toml:project.version"]
branch = "main"
upload_to_release = true

[tool.semantic_release.changelog]
exclude_commit_patterns = []

[tool.semantic_release.changelog.default_templates]
output_format = "md"
changelog_file = "CHANGELOG.md"

[tool.semantic_release.commit_parser_options]
allowed_tags = ["build", "chore", "ci", "docs", "feat", "fix", "perf", "refactor", "style", "test"]
minor_tags = ["feat"]
patch_tags = ["fix", "perf"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4", "SIM"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "--cov=src --cov-report=term-missing"
asyncio_mode = "auto"

[tool.coverage.run]
omit = ["tests/*", ".venv/*"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
mypy_path = "src"

[dependency-groups]
dev = [
    "mypy>=1.13.0",
    "pre-commit>=4.0.1",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.2",
    "types-requests>=2.32.0",
    "types-aiofiles>=24.1.0",
]
